chore: Rename fmt to parsers
