+++ src/main/java/clanker/Clanker.java
@@ -13,8 +13,8 @@ import clanker.task.DeadlineTask;
 import clanker.task.EventTask;
 import clanker.task.Task;
 import clanker.task.TodoTask;
-import fmt.CommandParser;
 import javafx.util.Pair;
+import parsers.CommandParser;
 import serde.Serde;
 
 public class Clanker {
