+++ src/main/java/parsers/CommandParser.java
@@ -1,4 +1,4 @@
-package fmt;
+package parsers;
 
 import java.util.Arrays;
 import java.util.HashMap;
