+++ src/test/java/parsers/CommandParserTest.java
@@ -1,4 +1,4 @@
-package fmt;
+package parsers;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
