+++ src/test/java/parsers/DateTimeParserTest.java
@@ -1,4 +1,4 @@
-package fmt;
+package parsers;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
